cargo: update deps and fix breaking changes

+++ kernel/src/allocator/mod.rs
@@ -49,7 +49,7 @@ pub fn init_heap(
 ) -> Result<(), MapToError<Size4KiB>> {
     let page_range = {
         let heap_start = VirtAddr::new(HEAP_START as u64);
-        let heap_end = heap_start + HEAP_SIZE - 1_u64;
+        let heap_end = heap_start + HEAP_SIZE as u64 - 1_u64;
 
         let heap_start_page = Page::containing_address(heap_start);
         let heap_end_page = Page::containing_address(heap_end);
